+++ src/proof/alethe/alethe_printer.h
@@ -15,46 +15,15 @@
 
 #include "cvc5_private.h"
 
-#ifndef CVC4__PROOF__ALETHE_PROOF_PRINTER_H
-#define CVC4__PROOF__ALETHE_PROOF_PRINTER_H
+#ifndef CVC5__PROOF__ALETHE_PROOF_PRINTER_H
+#define CVC5__PROOF__ALETHE_PROOF_PRINTER_H
 
-#include "printer/let_binding.h"
+#include "proof/alethe/alethe_let_binding.h"
 #include "proof/proof_node.h"
 
 namespace cvc5::internal {
 
 namespace proof {
-
-/** The Alethe-specific let binder.
- *
- * Differently from the regular let binder, where all letified subterms are
- * replaced by a fresh variable, the Alethe let binder replaces the first
- * occurrence of a term n (first visited on a post-order traversal) by a fresh
- * variable whose name is "(! n :named v)", in which `v` is another fresh
- * variable. The subsequent occurrences of `n` are replaced by `v`.
- */
-class AletheLetBinding : public LetBinding
-{
- public:
-  AletheLetBinding(uint32_t thresh);
-
-  /**
-   * Convert n based on the state of the let binding.
-   *
-   * The conversion is done as summarized as above, but the name of the fresh
-   * variable `v` is prefixed by `prefix`.
-   *
-   * @param n The node to convert
-   * @param prefix The prefix of variables to convert
-   * @return the converted node.
-   */
-  Node convert(Node n, const std::string& prefix);
-
- private:
-  /** The set of terms that have already been "decleared", i.e., already had
-   * their first occurrence replaced. */
-  std::unordered_set<Node> d_declared;
-};
 
 /**
  * The Alethe printer, which prints proof nodes in a Alethe proof, according to
@@ -111,4 +80,4 @@ class AletheProofPrinter
 
 }  // namespace cvc5::internal
 
-#endif /* CVC4__PROOF__ALETHE_PROOF_PRINTER_H */
+#endif /* CVC5__PROOF__ALETHE_PROOF_PRINTER_H */
